fix_codecheck_20260511 update001 Signed-off-by heguangyao <heguangyao1@h-partners.com>

+++ plugins/common/libs/image/libgifplugin/src/gif_encoder.cpp
@@ -414,7 +414,7 @@ void InitColorSubdivMap(ColorSubdivMap* colorSubdivMap, int32_t colorSubdivMapSi
         colorSubdivMap[i].pixelNum = 0;
         colorSubdivMap[i].colorNum = 0;
     }
-    colorSubdivMap[0].pixelNum = ((long)width) * height;
+    colorSubdivMap[0].pixelNum = static_cast<float>(width) * height;
 }
 
 void InitForQuantize(ColorCoordinate *colorCoordinate, ColorSubdivMap* colorSubdivMap)
